+++ packages/express/src/test/mocks/jwtUserMocks.ts
@@ -5,6 +5,7 @@ export const userJwtUserMock: JWTUser = new JWTUser({
   employeeReference: "1",
   organizationReference: "test",
   roles: [],
+  roleReferences: [],
   permissions: "MA==",
   segmentReference: null,
 });
@@ -14,6 +15,7 @@ export const laraJwtUserMock: JWTUser = new JWTUser({
   employeeReference: "0",
   organizationReference: "lara",
   roles: [],
+  roleReferences: [],
   permissions: "MA==",
   segmentReference: null,
 });